Regex adicionado ao projeto

+++ src/ts/main.ts
@@ -4,6 +4,7 @@
 
 let modoCriacao = false;
 let modoPessoa = false;
+let confirmacaoForm = false;
 
 /*Configuração dos botões e elementos do HTML*/
 const buttonExbicao = document.getElementById('buttonExibicao') as HTMLButtonElement || null;
@@ -16,10 +17,108 @@ const displayGrafico = document.getElementById('criacaoGrafico') as HTMLButtonEl
 /*Criando e salvando o form*/
 const cadForm = document.getElementById('itemForm')  as HTMLFormElement || null;
 
+document.addEventListener("DOMContentLoaded", (e) => {
+    e.preventDefault();
+
+    const nomePessoa = document.getElementById("nomeEmpresa") as HTMLInputElement;
+    const emailPessoa = document.getElementById("emailEmpresa") as HTMLInputElement;
+    const estadoEmpresa = document.getElementById("estadoEmpresa") as HTMLInputElement;
+    const paisEmpresa = document.getElementById("paisEmpresa") as HTMLInputElement;
+    const cepEmpresa = document.getElementById("cepEmpresa") as HTMLInputElement;
+
+    let nomePessoaRegex = /^[a-zA-Z\s]+$/gi;
+    let emailPessoaRegex = /[A-Za-z0-9\._-]+@[A-Za-z\.]+[A-Za-z]{2,3}/gi;
+    let estadoEmpresaRegex = /^[a-zA-Z\s]+$/gi;
+    let paisEmpresaRegex = /^[a-zA-Z\s]+$/gi;
+    let cepEmpresaRegex = /^[0-9]{5}-[0-9]{3}$/gi;
+
+    nomePessoa.addEventListener("input", (e) => {
+        e.preventDefault();
+
+        if(nomePessoa.value === ""){
+            nomePessoa.style.borderColor = "";
+        }
+
+        if(nomePessoaRegex.test(nomePessoa.value.trim())){
+            nomePessoa.style.borderColor = "green";
+            confirmacaoForm = true;
+        }else{
+            nomePessoa.style.borderColor = "red";
+            confirmacaoForm = false;
+        }
+
+    })
+
+    emailPessoa.addEventListener("input", (e) => {
+        e.preventDefault();
+
+        if(emailPessoa.value === ""){
+            emailPessoa.style.borderColor = "";
+        }
+
+        if(emailPessoaRegex.test(emailPessoa.value.trim())){
+            emailPessoa.style.borderColor = "green";
+            confirmacaoForm = true;
+        }else{
+            emailPessoa.style.borderColor = "red";
+            confirmacaoForm = false;
+        }
+    })
+
+    estadoEmpresa.addEventListener("input", (e) => {
+        e.preventDefault();
+
+        if(estadoEmpresa.value === ""){
+            estadoEmpresa.style.borderColor = "";
+        }
+
+        if(estadoEmpresaRegex.test(estadoEmpresa.value.trim())){
+            estadoEmpresa.style.borderColor = "green";
+            confirmacaoForm = true;
+        }else{
+            estadoEmpresa.style.borderColor = "red";
+            confirmacaoForm = false;
+        }
+    })
+
+    paisEmpresa.addEventListener("input", (e) => {
+        e.preventDefault();
+
+        if(paisEmpresa.value === ""){
+            paisEmpresa.style.borderColor = "";
+        }
+
+        if(paisEmpresaRegex.test(paisEmpresa.value.trim())){
+            paisEmpresa.style.borderColor = "green";
+            confirmacaoForm = true;
+        }else{
+            paisEmpresa.style.borderColor = "red";
+            confirmacaoForm = false;
+        }
+    })
+
+    cepEmpresa.addEventListener("input", (e) => {
+        e.preventDefault();
+
+        if(cepEmpresa.value === ""){
+            cepEmpresa.style.borderColor = "";
+        }
+
+        if(cepEmpresaRegex.test(cepEmpresa.value.trim())){
+            cepEmpresa.style.borderColor = "green";
+            confirmacaoForm = true;
+        }else{
+            cepEmpresa.style.borderColor = "red";
+            confirmacaoForm = false;
+        }
+    })
+})
+
+
 cadForm.addEventListener("submit", (e) =>{
     e.preventDefault();
 
-    //nao esta dando para adicionar o value direto no getElementById *-*
+    //Entradas do form
     const nomePessoa = document.getElementById("nomeEmpresa") as HTMLInputElement;
     const emailPessoa = document.getElementById("emailEmpresa") as HTMLInputElement;
     const idEmpresa = document.getElementById("idEmpresa") as HTMLInputElement;
@@ -30,7 +129,6 @@ cadForm.addEventListener("submit", (e) =>{
     const cepEmpresa = document.getElementById("cepEmpresa") as HTMLInputElement;
 
 
-
     let values: string[] = [];
 
     competencias.forEach(checkbox => {
@@ -38,6 +136,7 @@ cadForm.addEventListener("submit", (e) =>{
             values.push((checkbox as HTMLInputElement).value);
         }
     })
+
 
 
     const nome = nomePessoa.value;
@@ -61,27 +160,29 @@ cadForm.addEventListener("submit", (e) =>{
         adicionarEmpresaUsuario.textContent = `Adicionar ${tipo}`;
     }
 
-
-    if(modoPessoa){
-        //const empresa = new Empresa(nome, email, pais, id, estado, cep, descricao, checkCompetencias);
-        if(localStorage.hasOwnProperty("usuarios")){
-            usuarios = JSON.parse(localStorage.getItem("usuarios") || "[]");
+    if(confirmacaoForm){
+        if(modoPessoa){
+            //const empresa = new Empresa(nome, email, pais, id, estado, cep, descricao, checkCompetencias);
+            if(localStorage.hasOwnProperty("usuarios")){
+                usuarios = JSON.parse(localStorage.getItem("usuarios") || "[]");
+            }else{
+                usuarios = [];
+            }
+            if(!modoPessoa) return;
+            usuarios.push({nome, email, id, descricao, estado, checkCompetencias, pais, cep});
+            localStorage.setItem("usuarios", JSON.stringify(usuarios));
         }else{
-            usuarios = [];
+            if(localStorage.hasOwnProperty("empresas")){
+                empresas = JSON.parse(localStorage.getItem("empresas") || "[]");
+            }else{
+                empresas = [];
+            }
+            if(modoPessoa) return;
+            empresas.push({nome, email, id, descricao, estado, checkCompetencias, pais, cep});
+            localStorage.setItem("empresas", JSON.stringify(empresas));
         }
-        if(!modoPessoa) return;
-        usuarios.push({nome, email, id, descricao, estado, checkCompetencias, pais, cep});
-        localStorage.setItem("usuarios", JSON.stringify(usuarios));
-    }else{
-        if(localStorage.hasOwnProperty("empresas")){
-            empresas = JSON.parse(localStorage.getItem("empresas") || "[]");
-        }else{
-            empresas = [];
-        }
-        if(modoPessoa) return;
-        empresas.push({nome, email, id, descricao, estado, checkCompetencias, pais, cep});
-        localStorage.setItem("empresas", JSON.stringify(empresas));
     }
+
 
 
     nomePessoa.value = "";
@@ -214,7 +315,7 @@ function mudandoGrafico() {
 
     const labels = Object.keys(contagemCompetencias);
     const dados = Object.values(contagemCompetencias);
-
+    
     // Cria o gráfico
     return new Chart(canvas, {
         type: 'polarArea',
@@ -277,6 +378,8 @@ function iniciandoAplicacao() {
         });
     }
 }
+
+
 
 
 if (document.readyState === 'loading') {
